+++ clojure/src/tutkain/corerrb_vector/v0v1v1/clojure/core/rrb_vector/fork_join.clj
@@ -1,8 +1,0 @@
-(ns tutkain.corerrb-vector.v0v1v1.clojure.core.rrb-vector.fork-join
-  (:require [clojure.core.reducers :as r]))
-
-(def pool   @#'r/pool)
-(def task   @#'r/fjtask)
-(def invoke @#'r/fjinvoke)
-(def fork   @#'r/fjfork)
-(def join   @#'r/fjjoin)