+++ clojure/src/tutkain/nrepl/util/pprint.clj
@@ -1,12 +1,11 @@
 (ns tutkain.nrepl.util.pprint
   (:require
-   [tutkain.fipp.v0v6v23.fipp.edn :as fipp]))
+   [clojure.pprint :as pprint]))
 
 (defn useless?
   [entry]
   (let [^String s (-> entry first str)]
     (or
-      (.startsWith s "tutkain.")
       (.startsWith s "nrepl.")
       (.startsWith s "clojure.lang.AFn")
       (.startsWith s "clojure.lang.RestFn"))))
@@ -16,7 +15,7 @@
   (update (Throwable->map ex) :trace (partial remove useless?)))
 
 (defn pprint
-  [value writer options]
-  (fipp/pprint
+  [value writer _]
+  (pprint/pprint
     (if (instance? Throwable value) (humanize-exception value) value)
-    (assoc options :writer writer)))
+    writer))